combine test util into paginator test
